refactor(chat): extract content area and side effects from ChatScreen
Move the per-state content area (landing/loading/active, including the
comparison dual-pane and tab-pager layouts and buildComparisonDisplayMessages)
into ColumnScope.ChatContent in ChatContent.kt, and the screen's one-shot
LaunchedEffects/lifecycle/snackbar wiring into ChatScreenEffects.kt. ChatScreen
now delegates both, leaving its scaffold and dialog hosting.
