+++ feature/chat/src/androidMain/kotlin/com/garfiec/librechat/feature/chat/screen/ChatScreenEffects.kt
@@ -1,0 +1,113 @@
+package com.garfiec.librechat.feature.chat.screen
+
+import android.content.ClipData
+import android.content.ClipboardManager
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.compose.LifecycleEventEffect
+import com.garfiec.librechat.feature.chat.viewmodel.ChatUiState
+import com.garfiec.librechat.feature.chat.viewmodel.ChatViewModel
+
+/**
+ * Hosts the chat screen's one-shot side effects so [ChatScreen] reads as layout.
+ * Covers: new-conversation navigation handoff, error/share-link snackbars,
+ * fork/duplicate navigation, stream resume on foreground, the provider-key error
+ * snackbar, and back-navigation after a delete/archive clears the conversation.
+ */
+@Composable
+internal fun ChatScreenEffects(
+    uiState: ChatUiState,
+    shareLinkUrl: String?,
+    viewModel: ChatViewModel,
+    snackbarHostState: SnackbarHostState,
+    clipboardManager: ClipboardManager,
+    onConversationStart: ((String) -> Unit)?,
+    onNavigateToConversation: ((String) -> Unit)?,
+    onNavigateBack: (() -> Unit)?,
+    onNavigateToProviderKeys: (endpointName: String?) -> Unit,
+) {
+    // When a new conversation starts, navigate to Chat(conversationId) immediately
+    // (at StreamEvent.Created) so the NewChat landing page stays clean in the back
+    // stack. The new ChatViewModel at Chat(id) will resume the active stream.
+    // onPendingNavigationHandled() resets this ViewModel to a fresh landing state.
+    LaunchedEffect(uiState.pendingNavigationConversationId) {
+        val pendingId = uiState.pendingNavigationConversationId
+        if (pendingId != null && onConversationStart != null) {
+            onConversationStart(pendingId)
+            viewModel.onPendingNavigationHandled()
+        }
+    }
+
+    // Show errors in snackbar
+    LaunchedEffect(uiState.error) {
+        val error = uiState.error
+        if (error != null) {
+            snackbarHostState.showSnackbar(
+                message = error,
+                actionLabel = "Dismiss",
+            )
+            viewModel.dismissError()
+        }
+    }
+
+    UserKeyErrorSnackbarEffect(
+        viewModel = viewModel,
+        snackbarHostState = snackbarHostState,
+        onNavigateToProviderKeys = onNavigateToProviderKeys,
+    )
+
+    // Stream resume on foreground
+    LifecycleEventEffect(Lifecycle.Event.ON_PAUSE) { viewModel.onPause() }
+    LifecycleEventEffect(Lifecycle.Event.ON_RESUME) { viewModel.onResume() }
+
+    // Navigate to forked conversation
+    LaunchedEffect(uiState.forkedConversationId) {
+        val forkId = uiState.forkedConversationId
+        if (forkId != null) {
+            viewModel.onForkedConversationHandled()
+            if (onNavigateToConversation != null) {
+                onNavigateToConversation(forkId)
+            } else if (onConversationStart != null) {
+                onConversationStart(forkId)
+            }
+        }
+    }
+
+    // Navigate to duplicated conversation
+    LaunchedEffect(uiState.duplicatedConversationId) {
+        val dupId = uiState.duplicatedConversationId
+        if (dupId != null) {
+            viewModel.onDuplicatedConversationHandled()
+            if (onNavigateToConversation != null) {
+                onNavigateToConversation(dupId)
+            } else if (onConversationStart != null) {
+                onConversationStart(dupId)
+            }
+        }
+    }
+
+    // Copy share link to clipboard
+    LaunchedEffect(shareLinkUrl) {
+        val url = shareLinkUrl
+        if (url != null) {
+            clipboardManager.setPrimaryClip(ClipData.newPlainText("Share Link", url))
+            viewModel.onShareLinkHandled()
+            snackbarHostState.showSnackbar("Share link copied to clipboard")
+        }
+    }
+
+    // Navigate back after delete/archive (conversationId becomes null)
+    var hadConversation by remember { mutableStateOf(uiState.conversationId != null) }
+    LaunchedEffect(uiState.conversationId) {
+        if (hadConversation && uiState.conversationId == null) {
+            onNavigateBack?.invoke()
+        }
+        hadConversation = uiState.conversationId != null
+    }
+}
